+++ src/OpenLoco/Map/TileManager.h
@@ -25,6 +25,10 @@ namespace OpenLoco::Map::TileManager
     TileHeight getHeight(const map_pos& pos);
     void updateTilePointers();
     void reorganise();
+    map_pos screenGetMapXY(int16_t x, int16_t y);
+    uint16_t setMapSelectionTiles(int16_t x, int16_t y);
+    map_pos3 screenPosToMapPos(int16_t x, int16_t y);
+    uint16_t setMapSelectionSingleTile(int16_t x, int16_t y, bool setQuadrant = false);
     void mapInvalidateSelectionRect();
     void mapInvalidateTileFull(Map::map_pos pos);
     void mapInvalidateMapSelectionTiles();