+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 21.03+ (???)
 ------------------------------------------------------------------------
 - Feature: [#451] Optionally show an FPS counter at the top of the screen.
+- Feature: [#831] Add a tile inspector, allowing inspection of tile element data (read-only).
 - Feature: [#853] Allow unlocking FPS by detaching game logic from rendering.
 - Fix: [#804] Enter key not confirming save prompt.
 - Fix: [#809] Audio calculation not using the z axis.